Editorial revision

+++ calculator/src/compiler/interpreter.rs
@@ -57,13 +57,12 @@ mod tests {
 
     #[test]
     fn basics() {
-        assert_eq!(Interpreter::from_source("1").unwrap() as i32, 1);
-        assert_eq!(Interpreter::from_source("1 + 2").unwrap() as i32, 3);
-        // assert_eq!(Interpreter::source("(1 + 2)").unwrap() as i32, 3);
-        assert_eq!(Interpreter::from_source("2 + (2 - 1)").unwrap() as i32, 3);
-        assert_eq!(Interpreter::from_source("(2 + 3) - 1").unwrap() as i32, 4);
+        assert_eq!(Interpreter::from_source("1").unwrap(), 1);
+        assert_eq!(Interpreter::from_source("1 + 2").unwrap(), 3);
+        assert_eq!(Interpreter::from_source("2 + (2 - 1)").unwrap(), 3);
+        assert_eq!(Interpreter::from_source("(2 + 3) - 1").unwrap(), 4);
         assert_eq!(
-            Interpreter::from_source("1 + ((2 + 3) - (2 + 3))").unwrap() as i32,
+            Interpreter::from_source("1 + ((2 + 3) - (2 + 3))").unwrap(),
             1
         );
     }
